US20884 - [Continued] VERACODE - ROI Client - CWE-73 - External Control of File Name or Path

+++ src/client/McK.EIG.ROI.Client/Base/View/ROIPane.cs
@@ -257,7 +257,7 @@ namespace McK.EIG.ROI.Client.Base.View
                 {
                     try
                     {
-                        if (Validator.Validate(filePath, ROIConstants.FilepathValidation))
+                        if (Validator.Validate(filePath, ROIConstants.FilepathValidation)  && filePath.StartsWith(Environment.CurrentDirectory))
                         {
                             File.Delete(filePath);
                         }
